messagehub2elasticsearch/purge_old_indices/requirements.txt
schedule==0.4.3
elasticsearch==5.4.0
elasticsearch-curator==5.2.0
